+++ Editor/Scripts/ScriptTemplates/CreateNewMonoBehaviour.cs
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2025 Kokowolo. All Rights Reserved. 
+ * Copyright (c) 2026 Kokowolo. All Rights Reserved. 
  * Author(s): Kokowolo, Will Lacey
  * Date Created: April 26, 2022
  *